Stop storing a copy of the optimized Graph in FunctionLibraryRuntimeImpl::Item.
We currently only require a copy of the optimized Graph in order to implement FunctionLibraryRuntime::DebugString(). This change modifies the default behavior of FunctionLibraryRuntimeImpl::DebugString() to return a serialization of the unoptimized graph, and adds an option to InstantiateOptions that enables the optimized graph to be stored for debug purposes.

PiperOrigin-RevId: 272870225