Rewrite `TensorListResize` op into a if-branch to mimic the behavior of TF op.
TensorListResize will shrink/expand the input tensorlist based on the input size and requested size. If the requested size is larger than input tensorlist size, we need to concat the input tensorlist with an extended part initialized with 0. Otherwise, we need to slice 'size' rows from the input tensorlist and return.

PiperOrigin-RevId: 273315743