Allow tape.watched_variables() to be called after tape.gradient(..).
Previously a NoneType exception was thrown for the following code block
(calling gradient sets `tape._tape` to None if the tape is not persistent):

    grads = tape.gradient(loss, tape.watched_variables())
    optimizer.apply(grads, tape.watched_variables())

By keeping a reference to the watched variables whenever they are accessed or
when the tape is cleared by gradient(..) we allow the user to repeatedly call
the method at the expense of caching a tuple of `tf.Variable`s. Since these
variables are likely to be referenced elsewhere in the program this is unlikely
to cause a memory leak.

PiperOrigin-RevId: 274006599