Add an error handler to mlir compiler.
Add an EmissionContext class that registers a diagnostic handler and collects all the errors as a map from HloInstruction* to a vector of string representations of all the errors that occurred at that hlo instruction.
It also takes a callback that handles all those errors and is executed when the instance gets destroyed.

Also, add an InjectErrorsForTestingPass that emits errors from all operations and is used for testing.

PiperOrigin-RevId: 274155752